requirements.txt
gensim == 0.12.2
nltk == 3.0.2
numpy == 1.10.4
pandas == 0.16.2
regex == 2015.5.7
textblob == 0.9.0
tictacs == 0.0.3
xgboost == 0.4a30
BeautifulSoup == 3.2.1
scikit_learn == 0.17.1
PyYAML == 3.11
